requirements.txt
Django==5.2.6
djangorestframework==3.16.1
ollama==0.5.3
pytesseract==0.3.13
Pillow
requests
